Add func SwaggerNullableParameterFilter

+++ src/TeduBlog.Core/Models/Content/PostInListDto.cs
@@ -26,10 +26,10 @@ namespace TeduBlog.Core.Models.Content
 
         public PostStatus Status { set; get; }
         public bool IsPaid { get; set; }
-        public double RoyaltyAmount { get; set; }
+        public decimal RoyaltyAmount { get; set; }
         public DateTime? PaidDate { get; set; }
         public class AutoMapperProfiles : Profile
-        {
+        { 
             public AutoMapperProfiles()
             {
                 CreateMap<Post, PostInListDto>();
